added populateCrimeOption method
Gets all crimes from getCrimeData, then iterates over the json array and puts all the crimetype into a set. It is put in a set to make sure there no duplicates. Then unshifts an empty string to make sure it is the first element as before. Uses useState to make sure the optionCime lisr is reachable